+++ Models/BikeNews.cs
@@ -2,4 +2,4 @@
 
 namespace BikeServiceAPI.Models;
 
-public record BikeNews(int Id, BikeNewsType Type, string PictureLink, string Description);
+public record BikeNews(long Id, BikeNewsType Type, string PictureLink, string Description);
