+++ Models/Colleague.cs
@@ -5,7 +5,7 @@ namespace BikeServiceAPI.Models;
 public class Colleague : Person
 {
     public SkillLevel SkillLevel { get; private set; }
-    public Colleague(int id, string name, string email, string password, string phone, SkillLevel skillLevel, string? introduction = null) : base(id, name, email, password, phone, introduction)
+    public Colleague(long id, string name, string email, string password, string phone, SkillLevel skillLevel, string? introduction = null) : base(id, name, email, password, phone, introduction)
     {
         SkillLevel = skillLevel;
     }
